pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools",
    "packaging>=21.3",
    "twine",
    "vermin",
]

[project]
name = "edq-quizgen-simple-gui"
version = "0.0.1"
description = "A simple GUI for the Quiz Generator meant to be run on a local server."
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}

authors = [
    {name = "Eriq Augustine", email = "eaugusti@ucsc.edu"},
]

keywords = ['education']
classifiers = [
    'Intended Audience :: Education',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3.9',
]

dependencies = [
    "eq-quizgen>=0.2.8",
]

[project.urls]
Homepage = "https://github.com/edulinq/quizgen-simple-gui"
Repository = "https://github.com/edulinq/quizgen-simple-gui"

[tool.setuptools.packages.find]
where = ["."]
include = ["qgg*", "static"]
namespaces = false
